# api/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Tekton Hub
  description: HTTP services for managing Tekton Hub
  version: "1.0"
host: api.hub.tekton.dev
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /resource/{catalog}/{kind}/{name}:
    get:
      tags:
      - resource
      summary: ByCatalogKindName resource
      description: Find resources using name of catalog, resource name and kind of
        resource
      operationId: resource#ByCatalogKindName
      parameters:
      - name: pipelinesversion
        in: query
        description: To find resource compatible with a Tekton pipelines version,
          use this param
        required: false
        type: string
        pattern: ^\d+(?:\.\d+){0,2}$
      - name: catalog
        in: path
        description: name of catalog
        required: true
        type: string
      - name: kind
        in: path
        description: kind of resource
        required: true
        type: string
        enum:
        - task
        - pipeline
      - name: name
        in: path
        description: Name of resource
        required: true
        type: string
      responses:
        "302":
          description: Found response.
          headers:
            location:
              description: Redirect URL
              type: string
              format: uri
      schemes:
      - https
  /resource/{catalog}/{kind}/{name}/{version}:
    get:
      tags:
      - resource
      summary: ByCatalogKindNameVersion resource
      description: Find resource using name of catalog & name, kind and version of
        resource
      operationId: resource#ByCatalogKindNameVersion
      parameters:
      - name: catalog
        in: path
        description: name of catalog
        required: true
        type: string
      - name: kind
        in: path
        description: kind of resource
        required: true
        type: string
        enum:
        - task
        - pipeline
      - name: name
        in: path
        description: name of resource
        required: true
        type: string
      - name: version
        in: path
        description: version of resource
        required: true
        type: string
      responses:
        "302":
          description: Found response.
          headers:
            location:
              description: Redirect URL
              type: string
              format: uri
      schemes:
      - https
  /resource/{catalog}/{kind}/{name}/{version}/readme:
    get:
      tags:
      - resource
      summary: ByCatalogKindNameVersionReadme resource
      description: Find resource README using name of catalog & name, kind and version
        of resource
      operationId: resource#ByCatalogKindNameVersionReadme
      parameters:
      - name: catalog
        in: path
        description: name of catalog
        required: true
        type: string
      - name: kind
        in: path
        description: kind of resource
        required: true
        type: string
        enum:
        - task
        - pipeline
      - name: name
        in: path
        description: name of resource
        required: true
        type: string
      - name: version
        in: path
        description: version of resource
        required: true
        type: string
      responses:
        "302":
          description: Found response.
          headers:
            location:
              description: Redirect URL
              type: string
              format: uri
      schemes:
      - https
  /resource/{catalog}/{kind}/{name}/{version}/yaml:
    get:
      tags:
      - resource
      summary: ByCatalogKindNameVersionYaml resource
      description: Find resource README using name of catalog & name, kind and version
        of resource
      operationId: resource#ByCatalogKindNameVersionYaml
      parameters:
      - name: catalog
        in: path
        description: name of catalog
        required: true
        type: string
      - name: kind
        in: path
        description: kind of resource
        required: true
        type: string
        enum:
        - task
        - pipeline
      - name: name
        in: path
        description: name of resource
        required: true
        type: string
      - name: version
        in: path
        description: version of resource
        required: true
        type: string
      responses:
        "302":
          description: Found response.
          headers:
            location:
              description: Redirect URL
              type: string
              format: uri
      schemes:
      - https
...
  /resource/{id}/versions:
    get:
      tags:
      - resource
      summary: VersionsByID resource
      description: Find all versions of a resource by its id
      operationId: resource#VersionsByID
      parameters:
      - name: id
        in: path
        description: ID of a resource
        required: true
        type: integer
      responses:
        "302":
          description: Found response.
          headers:
            location:
              description: Redirect URL
              type: string
              format: uri
      schemes:
      - https
  /resource/version/{versionID}:
    get:
      tags:
      - resource
      summary: ByVersionId resource
      description: Find a resource using its version's id
      operationId: resource#ByVersionId
      parameters:
      - name: versionID
        in: path
        description: Version ID of a resource's version
        required: true
        type: integer
      responses:
        "302":
          description: Found response.
          headers:
            location:
              description: Redirect URL
              type: string
              format: uri
      schemes:
      - https
...
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: |-
      Secures endpoint by requiring a valid JWT retrieved via the /auth/login endpoint.

      **Security Scopes**:
        * `rating:read`: Read-only access to rating
        * `rating:write`: Read and write access to rating
        * `agent:create`: Access to create or update an agent
        * `catalog:refresh`: Access to refresh catalog
        * `config:refresh`: Access to refresh config file
        * `refresh:token`: Access to refresh user access token
    name: Authorization
    in: header